Added module (#1)
<!-- markdownlint-disable-next-line first-line-heading -->
## Description

Terraform Module that creates a SNS topic for budget alarms and a SNS topic for cloudwatch alarms. A lambda is created and subscribed to the two SNS topics. The lambda will send messages to teams based on the alarms.

## Type of changes

- [ ] Refactoring (non-breaking change)
- [x ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would change existing
functionality)
- [ ] Bug fix (non-breaking change which fixes an issue)

## Checklist

<!-- Go over all the following points, and put an `x` in all the boxes
that apply. -->

- [ x] I am familiar with the [contributing
guidelines](../docs/CONTRIBUTING.md)
- [x ] I have followed the code style of the project
- [ x] I have added tests to cover my changes
- [ x] I have updated the documentation accordingly
- [ ] This PR is a result of pair or mob programming

---

## Sensitive Information Declaration

To ensure the utmost confidentiality and protect your and others
privacy, we kindly ask you to NOT including [PII (Personal Identifiable
Information) / PID (Personal Identifiable
Data)](https://digital.nhs.uk/data-and-information/keeping-data-safe-and-benefitting-the-public)
or any other sensitive data in this PR (Pull Request) and the codebase
changes. We will remove any PR that do contain any sensitive
information. We really appreciate your cooperation in this matter.

- [ x] I confirm that neither PII/PID nor sensitive data are included in
this PR and the codebase changes.